refactor: update paper component
