+++ app/(tabs)/index.tsx
@@ -8,7 +8,7 @@ import { SafeAreaView } from "react-native-safe-area-context";
 
 import { queryClient } from "../_layout";
 
-import { Paper } from "@/components/shadow-view";
+import { Paper } from "@/components/paper";
 import { db } from "@/db/client";
 import { Task, tasks } from "@/db/schema";
 import { theme } from "@/styles/theme";
